refactor(core): absorb remaining @verdaccio/utils helpers and remove the package
Move the author/gravatar helpers into a new authorUtils export and getLatestVersion into pkgUtils, then delete the deprecated @verdaccio/utils package now that every symbol it exposed lives in @verdaccio/core.
