+++ packages/core/core/package.json
@@ -59,6 +59,7 @@
     "ajv": "8.20.0",
     "http-errors": "2.0.1",
     "http-status-codes": "2.3.0",
+    "lodash": "4.18.1",
     "minimatch": "10.2.6",
     "process-warning": "1.0.0",
     "semver": "7.8.5"
