--- packages/utils/.npmignore
@@ -1,6 +1,0 @@
-/*
-!/bin/**/*
-!/build/**/*
-!index.js
-!LICENSE
-!README.md
