--- packages/utils/package.json
@@ -1,70 +1,0 @@
-{
-  "name": "@verdaccio/utils",
-  "version": "8.2.2",
-  "description": "Verdaccio Utilities",
-  "keywords": [
-    "enterprise",
-    "modules",
-    "package",
-    "private",
-    "proxy",
-    "registry",
-    "repository",
-    "server",
-    "verdaccio"
-  ],
-  "homepage": "https://verdaccio.org",
-  "bugs": {
-    "url": "https://github.com/verdaccio/verdaccio/issues"
-  },
-  "license": "MIT",
-  "author": {
-    "name": "Juan Picado",
-    "email": "juanpicado19@gmail.com"
-  },
-  "repository": {
-    "type": "https",
-    "url": "https://github.com/verdaccio/verdaccio",
-    "directory": "packages/utils"
-  },
-  "funding": {
-    "type": "opencollective",
-    "url": "https://opencollective.com/verdaccio"
-  },
-  "main": "./build/index.js",
-  "module": "./build/index.mjs",
-  "types": "./build/index.d.ts",
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./build/index.d.ts",
-        "default": "./build/index.mjs"
-      },
-      "require": {
-        "types": "./build/index.d.ts",
-        "default": "./build/index.js"
-      }
-    },
-    "./build/*": "./build/*"
-  },
-  "scripts": {
-    "clean": "rimraf ./build",
-    "test": "cross-env NODE_ENV=test BABEL_ENV=test vitest run",
-    "type-check": "tsc --noEmit -p tsconfig.build.json",
-    "watch": "vite build --watch",
-    "build": "vite build && tsc --emitDeclarationOnly -p tsconfig.build.json"
-  },
-  "dependencies": {
-    "@verdaccio/core": "workspace:8.2.2",
-    "lodash": "4.18.1",
-    "minimatch": "10.2.6"
-  },
-  "devDependencies": {
-    "@verdaccio/types": "workspace:13.0.6",
-    "vite": "8.1.5",
-    "vitest": "4.1.10"
-  },
-  "engines": {
-    "node": ">=22"
-  }
-}
