--- packages/utils/tsconfig.build.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./build"
-  },
-  "include": ["src/**/*.ts"]
-}
