--- packages/utils/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "extends": "../../tsconfig.reference.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./build",
-    "preserveSymlinks": true,
-    "composite": true,
-    "declaration": true,
-    "declarationMap": true
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["src/**/*.test.ts"],
-  "references": [
-    {
-      "path": "../core/core"
-    },
-    {
-      "path": "../core/types"
-    }
-  ]
-}
