--- packages/utils/vite.config.mjs
@@ -1,3 +1,0 @@
-import { createLibConfig } from '../../vite.lib.config.mjs';
-
-export default createLibConfig(import.meta.dirname);
